Add priest

+++ Assets/Scripts/Entity/StatusEffect/Buff/Heal.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewHeal", menuName = "StatusEffect/Heal")]
 public class Heal : StatusEffect
 {
-    public int CasterATK;
+    public float CasterATK;
     public int percent;
 
     public override void OnApply(CharacterInBattle target)
     {
-        int Amount = CasterATK * (percent / 100);  
+        float Amount = CasterATK * (percent / 100f);  
         target.Heal(Amount);
     }
     public override void OnTurn(CharacterInBattle target)
